requirements.txt
# Versões fixas para maior reprodutibilidade (ajuste se necessário)
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
yfinance==0.2.40
PyPortfolioOpt==1.5.6
gymnasium==0.29.1
stable-baselines3==2.3.2
torch==2.2.2
matplotlib==3.8.4
streamlit==1.37.1
pyyaml==6.0.1
